requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain_community
langchain.core
langchain.groq
groq
python.dotenv